Bring the tree clean under ruff 0.16's default rules
149 findings were fixed by `ruff check --fix`, and a further 22 by the
unsafe fixes for RUF046, C408, RUF059, SIM117, SIM118 and SIM102. Those
are mechanical: PEP 585/604 annotations, import sorting, `dict()` calls
to literals and the like. RUF046 was checked rather than assumed —
`round()` returns a real `int` for `np.float64` as well as `float`, so
dropping the `int()` wrapper in tile_builder's pixel arithmetic cannot
leak a numpy scalar into an array index.

Eight sites needed a judgment call:

- d_train_cnn: `raise Exception` for an unknown device becomes ValueError.
- e_predict_json: the temp NDJSON file has to outlive its block, so
  `NamedTemporaryFile(delete=False)` + `close()` becomes `mkstemp`, which
  is the API for exactly that.
- manual_eval, pipelines/app: nested `with` statements combined. The
  manual_eval hunk is large but `git diff -w` reduces it to 8 lines — the
  rest is re-indentation.
- simple_cnn: ALLOWED_KWARGS annotated ClassVar; the deliberate `x = x`
  that keeps the trained checkpoint loadable gets a noqa rather than a
  rewrite.
- manifest_writer: PYI034 wants `Self`, which typing only gained in 3.11
  against a 3.10 floor, so the annotation stays and is suppressed.
- reference_data: TRY004 asks for TypeError, but every other rejection in
  that function raises ValueError and callers catch config errors as one
  kind, so the contract wins.

Two files needed reformatting afterwards: the FURB122 `writelines` fix
left a long line, and target-version py310 lets the formatter use
parenthesized context managers in reference_data.

Verified: `ruff check .` clean (and clean under --extend-select RUF100,
so every noqa is live), `ruff format --check .` clean, 129 tests pass,
compileall clean.

+++ displacement_tracker/util/tile_builder.py
@@ -38,7 +38,7 @@ def _compute_valid_fraction(data: np.ndarray, nodata) -> float:
 def tile_pixel_size(src, span_m: float) -> int:
     """Square tile side in pixels given a metric span and raster resolution."""
     px = abs(src.transform.a)
-    return int(round(span_m / px))
+    return round(span_m / px)
 
 
 def compute_tile_window(
@@ -98,8 +98,8 @@ def world_window(src, x: float, y: float, core_m: float, margin_m: float):
             return None
 
         col_f, row_f = (~src.transform) * (x, y)
-        r_center = int(round(row_f))
-        c_center = int(round(col_f))
+        r_center = round(row_f)
+        c_center = round(col_f)
 
         half = tile_px // 2
         r0 = max(0, min(r_center - half, src.height - tile_px))
@@ -127,8 +127,8 @@ def create_label_from_feats(
 
     for feat in feats:
         feat_lon, feat_lat = feat["geometry"]["coordinates"]
-        local_col = int(round((feat_lon - lon_min) / lon_span * (w - 1)))
-        local_row = int(round((lat_max - feat_lat) / lat_span * (h - 1)))
+        local_col = round((feat_lon - lon_min) / lon_span * (w - 1))
+        local_row = round((lat_max - feat_lat) / lat_span * (h - 1))
 
         for dr in (-1, 0, 1):
             for dc in (-1, 0, 1):
@@ -163,8 +163,8 @@ def _read_prewar_tile(
         lat_c = 0.5 * (lat_min + lat_max)
         pxs, pys = transform("EPSG:4326", prewar_src.crs, [lon_c], [lat_c])
         col_f, row_f = (~prewar_src.transform) * (pxs[0], pys[0])
-        r_center = int(round(row_f))
-        c_center = int(round(col_f))
+        r_center = round(row_f)
+        c_center = round(col_f)
 
         half = tile_px // 2
         rpre0 = max(0, min(r_center - half, prewar_src.height - tile_px))
